requirements.txt
GitPython==3.1.27
jupyter==1.0.0
nltk==3.8.1
pandas==1.4.3
plotly==5.9.0
kaleido==0.2.1
